requirements.txt
psycopg2-binary
faker
redis